feat - conectando back end com o banco de dados e gerando as tabelas com o flyway

+++ src/main/java/github/MaxBr221/dtos/usuario/ClienteFrequencia.java
@@ -1,6 +1,6 @@
 package github.MaxBr221.dtos.usuario;
 
-public record ClienteFrequencia(Long idCliente, String nome, int totalAgendamentos) {
+public record ClienteFrequencia(Long idCliente, String nome, Long totalAgendamentos) {
     public ClienteFrequencia(ClienteFrequencia usuarios) {
         this(usuarios.idCliente,
                 usuarios.nome,
